+++ .eslintignore
@@ -1,0 +1,2 @@
+# Файлы сборки webpack
+webpack.config.js
